Moved challenge settings form to components
