+++ src/components/challengeSettingsForm/challengeDateInputs/challengeDateInputs.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { Columns, InputField } from "../../../../components/components";
+import { Columns, InputField } from "../../components";
 
 const ChallengeDateInputs = ({
   startDate,
