+++ src/components/challengeSettingsForm/challengeTitleInput/challengeTitleInput.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { InputField } from "../../../../components/components";
+import { InputField } from "../../components";
 
 import { getTitlePlaceholder } from "./challengeTitleInput.utility";
