+++ src/components/challengeSettingsForm/challengeTypeSelector/challengeTypeSelector.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { RadioButtonGroup } from "../../../../components/components";
+import { RadioButtonGroup } from "../../components";
 
 import { getTypeOptions } from "./challengeTypeSelector.utility";
