+++ src/components/components.js
@@ -1,15 +1,17 @@
-import Button from './button/button'
-import Chrome from './chrome/chrome'
-import Columns from './columns/columns'
-import ConfirmationDialog from './confirmationDialog/confirmationDialog'
-import Gap from './gap/gap'
-import InputField from './inputField/inputField'
-import Modal from './modal/modal'
-import RadioButtonGroup from './radioButtonGroup/radioButtonGroup'
-import Select from './select/select'
+import Button from "./button/button";
+import ChallengeSettingsForm from "./challengeSettingsForm/challengeSettingsForm";
+import Chrome from "./chrome/chrome";
+import Columns from "./columns/columns";
+import ConfirmationDialog from "./confirmationDialog/confirmationDialog";
+import Gap from "./gap/gap";
+import InputField from "./inputField/inputField";
+import Modal from "./modal/modal";
+import RadioButtonGroup from "./radioButtonGroup/radioButtonGroup";
+import Select from "./select/select";
 
 export {
   Button,
+  ChallengeSettingsForm,
   Chrome,
   Columns,
   ConfirmationDialog,
@@ -17,5 +19,5 @@ export {
   InputField,
   Modal,
   RadioButtonGroup,
-  Select
-}
+  Select,
+};
